+++ src/lib_base/protocol.ml
@@ -46,6 +46,7 @@ and env_version =
   | V11
   | V12
   | V13
+  | V14
 
 let compare_version = Stdlib.compare
 
@@ -80,6 +81,7 @@ let module_name_of_env_version = function
   | V11 -> "V11"
   | V12 -> "V12"
   | V13 -> "V13"
+  | V14 -> "V14"
 
 let env_version_encoding =
   let open Data_encoding in
@@ -99,7 +101,8 @@ let env_version_encoding =
          | V10 -> 10
          | V11 -> 11
          | V12 -> 12
-         | V13 -> 13)
+         | V13 -> 13
+         | V14 -> 14)
        (function
          | 0 -> V0
          | 1 -> V1
@@ -115,6 +118,7 @@ let env_version_encoding =
          | 11 -> V11
          | 12 -> V12
          | 13 -> V13
+         | 14 -> V14
          | x -> failwith (Format.sprintf "unexpected environment version: %d" x))
        uint16