+++ src/lib_protocol_compiler/registerer/tezos_protocol_registerer.mli
@@ -79,6 +79,10 @@ module type PROTOCOL_V13 = functor
   (Env : Tezos_protocol_environment_sigs.V13.T)
   -> Env.Updater.PROTOCOL
 
+module type PROTOCOL_V14 = functor
+  (Env : Tezos_protocol_environment_sigs.V14.T)
+  -> Env.Updater.PROTOCOL
+
 type proto_env =
   | V0 of (module PROTOCOL_V0)
   | V1 of (module PROTOCOL_V1)
@@ -94,6 +98,7 @@ type proto_env =
   | V11 of (module PROTOCOL_V11)
   | V12 of (module PROTOCOL_V12)
   | V13 of (module PROTOCOL_V13)
+  | V14 of (module PROTOCOL_V14)
 
 val register : string -> proto_env -> unit