+++ CHANGES.rst
@@ -54,6 +54,8 @@ Proxy Server
 Protocol Compiler And Environment
 ---------------------------------
 
+- Added a new version of the protocol environment (V14). (MR :gl:`!15345`)
+
 Codec
 -----